+++ src/Layers/xrRender/FSkinned.cpp
@@ -426,7 +426,7 @@ void CSkeletonX_ext::_Load_hw	(Fvisual& V, void *	_verts_)
 				Vertices1W.create		( crc, V.vCount, (vertBoned1W*)_verts_);
 			}
 
-			u32 vStride = ComputeVertexSize(dwDecl_01W, 0);
+			u32 vStride = (u32)ComputeVertexSize(dwDecl_01W, 0);
 			VERIFY	(vStride==sizeof(vertHW_1W));
 //			BYTE*	bytes		= 0;
 			VERIFY				(NULL==V.p_rm_Vertices);
@@ -467,7 +467,7 @@ void CSkeletonX_ext::_Load_hw	(Fvisual& V, void *	_verts_)
 				Vertices2W.create		( crc, V.vCount, (vertBoned2W*)_verts_);
 			}
 
-			u32 vStride = ComputeVertexSize(dwDecl_2W, 0);
+			u32 vStride = (u32)ComputeVertexSize(dwDecl_2W, 0);
 			VERIFY				(vStride==sizeof(vertHW_2W));
 //			BYTE* bytes			= 0;
 			VERIFY				(NULL==V.p_rm_Vertices);
@@ -508,7 +508,7 @@ void CSkeletonX_ext::_Load_hw	(Fvisual& V, void *	_verts_)
 				Vertices3W.create		( crc, V.vCount, (vertBoned3W*)_verts_);
 			}
 
-			u32 vStride = ComputeVertexSize(dwDecl_3W, 0);
+			u32 vStride = (u32)ComputeVertexSize(dwDecl_3W, 0);
 
 			VERIFY					(vStride==sizeof(vertHW_3W));
 //			BYTE*	bytes			= 0;
@@ -553,7 +553,7 @@ void CSkeletonX_ext::_Load_hw	(Fvisual& V, void *	_verts_)
 				Vertices4W.create		( crc, V.vCount, (vertBoned4W*)_verts_);
 			}
 
-			u32 vStride = ComputeVertexSize(dwDecl_4W, 0);
+			u32 vStride = (u32)ComputeVertexSize(dwDecl_4W, 0);
 			VERIFY					(vStride==sizeof(vertHW_4W));
 //			BYTE*	bytes			= 0;
 			VERIFY					(NULL==V.p_rm_Vertices);
