+++ src/Layers/xrRender/SkeletonCustom.h
@@ -66,7 +66,7 @@ public:
 #endif
 
 	IC CKinematics*		Parent				(){return m_Parent;}
-	IC u32				VCount				(){return m_Faces.size()*3;}
+	IC u32				VCount				(){return (u32)m_Faces.size()*3;}
 	IC bool				Similar				(ref_shader& sh, const Fvector& cp, float eps){return (m_Shader==sh)&&m_ContactPoint.similar(cp,eps);}
 	IC float			TimeStart			(){return m_fTimeStart;}
 	IC const Fmatrix*	XFORM				(){return m_XForm;}
