+++ src/xrEngine/XR_IOConsole.cpp
@@ -323,7 +323,7 @@ void CConsole::OnRender()
 	}
 	
 	// ---------------------
-	u32 log_line = LogFile->size()-1;
+	u32 log_line = (u32)LogFile->size()-1;
 	ypos -= LDIST;
 	for(int i = log_line - scroll_delta; i >= 0; --i) {
 		ypos -= LDIST;
@@ -454,7 +454,7 @@ void CConsole::DrawBackgrounds(bool bGame) {
 
 	// --------------------------- scroll bar --------------------
 
-	u32 tips_sz = m_tips.size();
+	u32 tips_sz = (u32)m_tips.size();
 	if (tips_sz > VIEW_TIPS_COUNT) {
 		Frect rb, rs;
 		
@@ -466,7 +466,7 @@ void CConsole::DrawBackgrounds(bool bGame) {
 
 		VERIFY( rb.y2 - rb.y1 >= 1.0f );
 		float back_height = rb.y2 - rb.y1;
-		float u_height = (back_height * VIEW_TIPS_COUNT)/ float(tips_sz);
+		float u_height = (back_height * (float)VIEW_TIPS_COUNT)/ float(tips_sz);
 		if (u_height < 0.5f * font_h) {
 			u_height = 0.5f * font_h;
 		}
@@ -688,7 +688,7 @@ IConsole_Command* CConsole::find_next_cmd( LPCSTR in_str, shared_str& out_str )
 }
 
 bool CConsole::add_next_cmds(LPCSTR in_str, vecTipsEx& out_v) {
-	u32 cur_count = out_v.size();
+	u32 cur_count = (u32)out_v.size();
 	if (cur_count >= MAX_TIPS_COUNT) {
 		return false;
 	}
@@ -726,7 +726,7 @@ bool CConsole::add_next_cmds(LPCSTR in_str, vecTipsEx& out_v) {
 
 bool CConsole::add_internal_cmds( LPCSTR in_str, vecTipsEx& out_v )
 {
-	u32 cur_count = out_v.size();
+	u32 cur_count = (u32)out_v.size();
 	if (cur_count >= MAX_TIPS_COUNT) {
 		return false;
 	}
@@ -866,7 +866,7 @@ void CConsole::select_for_filter(LPCSTR filter_str, vecTips& in_v, vecTipsEx& ou
 {
 	out_v.clear();
 
-	u32 in_count = in_v.size();
+	u32 in_count = (u32)in_v.size();
 	if (in_count == 0 || !filter_str) 
 	{
 		return;
