+++ src/xrEngine/editor_environment_ambients_effect_id.cpp
@@ -36,12 +36,12 @@ effect_id::~effect_id								()
 
 LPCSTR const* effect_id::collection					()
 {
-	return							(&*m_manager.effects_ids().begin());
+	return (&*m_manager.effects_ids().begin());
 }
 
 u32 effect_id::collection_size						()
 {
-	return							(m_manager.effects_ids().size());
+	return u32(m_manager.effects_ids().size());
 }
 
 void effect_id::fill								(editor::property_holder_collection* collection)
