+++ src/xrGame/alife_trader_abstract.cpp
@@ -215,7 +215,7 @@ void CSE_ALifeTraderAbstract::add_online	(const bool &update_registries)
 
 void add_offline_impl						(CSE_ALifeDynamicObject *object, const xr_vector<ALife::_OBJECT_ID> &saved_children, const bool &update_registries)
 {
-	for (u32 i=0, n=saved_children.size(); i<n; ++i) {
+	for (u32 i=0, n= (u32)saved_children.size(); i<n; ++i) {
 		CSE_ALifeDynamicObject	*child = smart_cast<CSE_ALifeDynamicObject*>(ai().alife().objects().object(saved_children[i],true));
 		R_ASSERT				(child);
 		child->m_bOnline		= false;
