+++ src/xrGame/obstacles_query.cpp
@@ -44,8 +44,8 @@ void obstacles_query::set_intersection	(const obstacles_query &query)
 void obstacles_query::merge				(const AREA &object_area)
 {
 	AREA temp(std::move(m_area));
-	u32 area_size = temp.size();
-	u32 destination_size = area_size + object_area.size();
+	size_t area_size = temp.size();
+	size_t destination_size = area_size + object_area.size();
 	m_area.resize(destination_size);
 	m_area.erase(
 		std::set_union(
